# Conflicts:
#	Source/ACE.Server/Factories/WorldObjectFactory.cs
#	Source/ACE.Server/WorldObjects/Creature_Death.cs
#	Source/ACE.Server/WorldObjects/Monster_Melee.cs
#	Source/ACE.Server/WorldObjects/Monster_Tick.cs

+++ Source/ACE.Server/WorldObjects/WorldObject_Equipment.cs
@@ -23,7 +23,7 @@ namespace ACE.Server.WorldObjects
                     wo.Shade = item.Shade;
 
                 if (item.StackSize > 0)
-                    wo.StackSize = (ushort)item.StackSize;
+                    wo.StackSize = item.StackSize;
 
                 items.Add(wo);
             }
